refactor(server): extract inbox routes to inboxRoutes.ts
Slice 3 of the server.ts split. Lifts the 2 inbox routes (`GET /inbox`
list, `GET /inbox/<filename>.md` read-by-filename) out of server.ts and
into a new module src/inboxRoutes.ts. Delegation pattern matches PRs #97
and #98:

    if (tryHandleInboxRoute(req, res, parsedUrl)) return;

Mechanical lift — handler bodies are byte-identical, including the
path-traversal check that rejects `/` and `\` in the filename.
Pre-extraction grep confirmed zero `this.` references; the routes only
reference `path`, `os`, and `fs/promises`.

server.ts: 2362 → 2266 lines (-96, another -4%); cumulative across all
three slice PRs (#97/#98/this) the original 1090+133+96 = 1319 lines of
contiguous boilerplate are now in three focused modules.

Verified:
- Full bridge test suite: 322 files / 4492 tests green (no regressions)
- Zero LSP diagnostics on both files (errors + warnings)
- Biome clean
