+++ src/server.ts
@@ -16,6 +16,7 @@ import {
 } from "./connectorRoutes.js";
 import { timingSafeStringEqual } from "./crypto.js";
 import { renderDashboardHtml } from "./dashboard.js";
+import { tryHandleInboxRoute } from "./inboxRoutes.js";
 import type { Logger } from "./logger.js";
 import type { OAuthServer } from "./oauth.js";
 import {
@@ -1002,112 +1003,10 @@ export class Server extends EventEmitter<ServerEvents> {
         return;
       }
 
-      // ── Inbox routes ────────────────────────────────────────────────────
-      if (parsedUrl.pathname === "/inbox" && req.method === "GET") {
-        void (async () => {
-          try {
-            const { readdir, readFile, stat } = await import(
-              "node:fs/promises"
-            );
-            const { existsSync } = await import("node:fs");
-            const inboxDir = path.join(os.homedir(), ".patchwork", "inbox");
-            if (!existsSync(inboxDir)) {
-              res.writeHead(200, { "Content-Type": "application/json" });
-              res.end(JSON.stringify({ items: [] }));
-              return;
-            }
-            const files = (await readdir(inboxDir)).filter((f) =>
-              f.endsWith(".md"),
-            );
-            const items = await Promise.all(
-              files.map(async (name) => {
-                const filePath = path.join(inboxDir, name);
-                const [content, stats] = await Promise.all([
-                  readFile(filePath, "utf8"),
-                  stat(filePath),
-                ]);
-                const stripped = content
-                  .split("\n")
-                  .filter((l) => !l.startsWith("#"))
-                  .join("\n")
-                  .trim();
-                return {
-                  name,
-                  path: filePath,
-                  modifiedAt: stats.mtime.toISOString(),
-                  preview: stripped.slice(0, 200),
-                };
-              }),
-            );
-            items.sort(
-              (a, b) =>
-                new Date(b.modifiedAt).getTime() -
-                new Date(a.modifiedAt).getTime(),
-            );
-            res.writeHead(200, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ items }));
-          } catch (err) {
-            res.writeHead(500, { "Content-Type": "application/json" });
-            res.end(
-              JSON.stringify({
-                error: err instanceof Error ? err.message : String(err),
-              }),
-            );
-          }
-        })();
+      // ── Inbox routes (extracted to src/inboxRoutes.ts) ───────────────────
+      if (tryHandleInboxRoute(req, res, parsedUrl)) {
         return;
       }
-
-      const inboxFileMatch = parsedUrl.pathname?.match(
-        /^\/inbox\/([^/]+\.md)$/,
-      );
-      if (inboxFileMatch && req.method === "GET") {
-        void (async () => {
-          try {
-            const { readFile, stat } = await import("node:fs/promises");
-            const filename = decodeURIComponent(inboxFileMatch[1] ?? "");
-            // Prevent path traversal — filename must not contain directory separators
-            if (filename.includes("/") || filename.includes("\\")) {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(JSON.stringify({ error: "Invalid filename" }));
-              return;
-            }
-            const filePath = path.join(
-              os.homedir(),
-              ".patchwork",
-              "inbox",
-              filename,
-            );
-            const [content, stats] = await Promise.all([
-              readFile(filePath, "utf8"),
-              stat(filePath),
-            ]);
-            res.writeHead(200, { "Content-Type": "application/json" });
-            res.end(
-              JSON.stringify({
-                name: filename,
-                content,
-                modifiedAt: stats.mtime.toISOString(),
-              }),
-            );
-          } catch (err: unknown) {
-            const code = (err as NodeJS.ErrnoException).code;
-            if (code === "ENOENT") {
-              res.writeHead(404, { "Content-Type": "application/json" });
-              res.end(JSON.stringify({ error: "Not found" }));
-            } else {
-              res.writeHead(500, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  error: err instanceof Error ? err.message : String(err),
-                }),
-              );
-            }
-          }
-        })();
-        return;
-      }
-      // ── End inbox routes ─────────────────────────────────────────────────
 
       const recipeNameRunMatch =
         req.method === "POST"
